+++ packages/describe-package/tsup.config.js
@@ -2,13 +2,6 @@ import baseConfig from '../../tsup.config';
 
 export default {
   ...baseConfig,
+  clean: false,
   bundle: true,
-  // shims: true,
-  // target: "chrome106",
-  platform: 'browser', // TODO I nee to create a browser build and export that differently
-  // external: ['node-fetch'],
-  // external: ['fs', 'events', 'stream', 'path', 'util', 'constants', 'assert'],
-  noExternal: ["@openfn", "typescript", "cross-fetch", "url-join"],
-  // target: 'chrome'
-
 };